WIP: Allow Experience CS to list all public projects
We currently display the list of Scratch projects in the Experience CS Admin
Dashboard which allows the learning team to author this content.

This is fine for the moment, however once teachers and students start to remix
these projects more and more projects will begin to appear on the admin
dashboard. The learning team only want to be dealing with the public projects
since these are the ones authored by them.

In order to solve this issue we have settled on modifying the existing
`ProjectsController#index` action so that if an `experience-cs-admin` calls the
controller action with the `all_public` query string set it will return a
paginated list of public projects. The client (in this case the `experience-cs`
app) can then filter out the Python and HTML projects.