Add administrative_area to SchoolDashboard
This change makes it easier to scan the list of recently added schools
and work out whether they are in a US State that is of particular
interest to the Experience CS customer success team. Before this
change, Kendall, who is part of the team, told us she had to click
into each school individually to find out this information.