Allow admin to search for a school in the NCES database
Our customer success team want to know which School District a school
in the US belongs to. At the moment we do not collect this information
during registration so the team have to spend time working this out
from the School's website and other information.

The Institute of Education Sciences maintains a website with data
about all public schools in the US, including their School District.

This commit adds a link to the bottom of the school show page in the
administrate dashboard to take the user to the search page on
`https://nces.ed.gov` for the school's ZIP code. I've decided to only
show this link for schools with the country code "US" to avoid
confusion. I've also decided to search for ZIP code rather than say,
school name. In my testing searching by school name works very well
when there's an exact match, but can return zero results when there's
a slight difference in spelling/abbreviations. Using the ZIP code
seems to reliably return the school (along with other schools close
by).

I've shown a rough version of this feature to the team who will use it
and they think it will be very useful until we start collecting this
data at source.

It's worth mentioning that the department that maintains this
data has been disbanded[1], so it may be the case that the website we
link to eventually goes away or becomes out-of-date. There appear to
be some efforts underway to preserve the data, but I've decided not to
pursue linking to those at this time.

[1] https://abcnews.go.com/Politics/education-department-cuts-agency-compiles-nations-report-card/story?id=119735831