Respond with JSON when ParameterMissing raised
Previously there was no body in the response which was inconsistent with
the approach taken in more specific API controller actions and made it
more awkward to implement a client for the API.

Now we respond with JSON including an error key containing a string made
up of the exception's class and message. This should make it easier to
implement a client for the API and easier to diagnose problems that
occur in an API request.