Respond with JSON when user is not authorised
Previously there was no body in the response which was inconsistent with
the approach taken in more specific API controller actions and made it
more awkward to implement a client for the API.

Now we respond with JSON including an error key containing an error
message. This should make it easier to implement a client for the API
and easier to diagnose problems that occur in an API request.