Respond with JSON when unhandled exception raised
Previously the HTML error template (`public/500.html`) was rendered as
the response which was inconsistent with the approach we're now taking
for other errors and made it more awkward to implement a client for the
API.

Now we respond with JSON including an error key containing a string made
up of the exception's class and message. This should make it easier to
implement a client for the API and easier to diagnose problems that
occur in an API request.

Note that the new `rescue_from` will not rescue unhandled exceptions
that occur in middleware, but we can tackle that separately if it turns
out to be a problem.

Also note that the new `rescue_from` ifor `StandardError` needs to be
defined first so that it's last in terms of matching exceptions in order
to get the fall-through behaviour that we want.

Previously an unhandled exception would've been reported to Sentry
automatically by the `Sentry::Rails::CaptureExceptions` middleware added
by sentry-rails, so to preserve that behaviour we explicitly call
`Sentry.capture_exception`.