Fix OperationResponse in ClassMember::Create
If there are any errors in the `#create_class_teachers` or
`#create_class_students` methods they are added to the
`OperationResponse` Hash under the `:errors` key. However,
`OperationResponse#success?` only takes the `:error` key into account.

Previously, it was possible for `OperationResponse#success?` to return
`true` even though one or more errors had been added to the `:errors`
key. This in turn meant that `Api::ClassMembersController#create` could
end up "thinking" the operation had been a success when in fact it had
(at least partially failed).

In such a scenario, the `:class_members` key on `OperationResponse`
could end up being an empty array. This then resulted in a
`NoMethodError` when those actions tried to render the JSON for the
class member(s) [1], because the `@class_member` ivar was `nil`.

Now, whenever an error is added to the `:errors` key, we also ensure
that a more generic error message is added to the `:error` key
referencing the `:errors` key. This means that
`OperationResponse#success?` now returns `false` when there are any
errors and the controller action returns an error response [2].

Note that a similar problem affected the
`Api::ClassMembersController#create_batch` action which has also been
fixed.

I did consider changing the implementation of
`OperationResponse#success?` to take the `:errors` key into account, but
it didn't really make sense to do that for a single use case. I
considered storing the Hash of errors against the `:error` key instead
of against the `:errors` key, but that would be less backwardly
compatible since most of the time the `:error` key just contains a
string error message. So I think this solution is probably the best
option for now.

I noticed this issue while spiking on
RaspberryPiFoundation/experience-cs#1152. I suspect the reason this
hasn't been a problem for the `ClassMembersPage` in `editor-standalone`,
because its UI effectively prevents the user from causing any validation
errors.

[1]: https://github.com/RaspberryPiFoundation/editor-api/blob/8a9012ca344bfa7976a83f024b48933a36139853/app/controllers/api/class_members_controller.rb#L39-L40
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/8a9012ca344bfa7976a83f024b48933a36139853/app/controllers/api/class_members_controller.rb#L42