Change the response format for ClassMembersController#create_batch
This preserves existing behaviour but changes the format of the
response to more easily support how we want to display this data in
the Code Editor for Education class memberships page.

TODO: are we??

We're fairly confident that this end point isn't used anywhere else,
so although this is a breaking change we're in control of the sole
client of this API so can manage the roll-out.