Change the response format for `ClassMembersController#create_batch`
So that it treats the request as successful even if not all (or any) of
the members could be added to the class. Previously if _any_ member
couldn't be added (likely due to them already being in the class) then
the response code was 422 and the response only contained information
about the errors and nothing about the members that were succesfully
added.

In order to use this `#create_batch` endpoint in Editor Standalone[1] we
need to differentiate between a request that was processed successfully
(irrespective of how many members were added to the class) and a
request that couldn't be processed (e.g. because none of the supplied
members exist in Profile API).

We've modified the return value from `#create_batch` such that it
contains enough information to differentiate between the members that
were added to the class and those that couldn't be added, along with the
reasons they couldn't be added.

In the case where none of the supplied members exist in Profile API
we've retained the existing behaviour of returning an `error` key only
in the response.

We think it's safe to make this change because the only place we'd
expect this endpoint to be used is in editor-standalone and we're
confident it's not currently being used in that project.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/pull/624

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>