Make UserJob track a Batch instead of a Job.
This commit changes UserJob from having a 1:1 relationship with
GoodJob::Job to tracking the ID of a GoodJob::Batch.

This means that the UserJobsController had to change to track the state
of a batch and emulate the status messages that a Job used to provide
(e.g. runnning, queued, etc.).

The front end can now poll for the status of this job.