Enable Bulk User Creation for Larger Batches of Users (#597)
## Status

This user story is in [Experience CS #1146 - Adding Students - Increase
Upload
Limit](https://github.com/RaspberryPiFoundation/experience-cs/issues/1146).

The overall goal of this PR is to allow Code Editor to ingest CSV
uploads of new users much larger than the current limit of 50 that is
imposed by Profile. However, it was also a goal NOT to change the 50-row
limit in the `createStudents` endpoint in Profile.

Therefore the high level strategy to solving this problem is:

1. Change Profile to allow fast validation of large batches of students
(done in [PR #1948 in
Profile](https://github.com/RaspberryPiFoundation/profile/pull/1948))
2. Have Code Editor accept a large batch of students and call that
endpoint to ensure the entire batch is valid.
3. Have Code Editor split the large batch into multiples of 50 students
and feed that into Profile - even if that takes substantial time to
complete.

This PR depends on the branch in [PR
#1948](https://github.com/RaspberryPiFoundation/profile/pull/1948) in
Profile and [PR #628 in
editor-standalone](https://github.com/RaspberryPiFoundation/editor-standalone/pull/628),
which removes the 50-row limit check on the front end.

## Points for consideration:

Please could reviewers focus on concurrency safety and any issues that
could arise if a batch takes a long time to run?

## What's changed?

- Changes `ProfileApiClient` to use the new `/preflight-student-upload`
endpoint in Profile.
- Removes batch validation from `SchoolStudent::CreateBatch` and instead
validates the entire upload using the new
`SchoolStudents::ValidateBatch` operation.
- Introduces a new GoodJob queue just for user creation.
- Uses a `GoodJob::Batch` to group individual `CreateStudentsJob` and
ensure that the enqueue of the entire batch of 50-user creation jobs is
atomic.
- Uses the `description` field of a `GoodJob::Batch` as a kind of
concurrency key, since batches don't have concurrency keys. The
controller will refuse to enqueue a new batch if there exists a batch
whose description matches the school ID and which has not been completed
or discarded.

## Steps to perform after deploying to production

This PR will require a DB migration due to adding a column to the
`user_jobs` table.