Add ImportSchoolsJob to track imports.
This commit introduces an ImportSchoolsJob type of GoodJob job to allow
for enqueueing imports and tracking their asynchronous completion.

The controller will return to the caller a Job ID which can be used to
track the progress of these jobs and retrieve their results when done.

The SchoolImportResult class tracks the job and provides access to the
results.

There are structured error codes for possible issues in
SchoolImportError.