Add a new Administrate page to allow and track uploads.
This commit adds new UI in /admin to allow users to upload CSVs of
schools and track the results of the upload jobs in the UI.

They can also download CSV files of the resulting school structures.

Add index, show and new views for Administrate

Add SchoolImportResultDashboard

Add routes for admin interface for school_import_results

Add SchoolImportResultsController to show results in the UI.

This commit also adds code to new.html.erb to show validation errors if
the uploaded CSV can't be enqueued as a batch.