update Project.with_user to not call profile api when user is a student (#627)
## Status

- Closes #596 

## What's changed?

Update `Project.with_user` to not call profile api when user is a
student as this call will always fail. Students cannot use that API at
the moment so this calls gets 500.

No outwards change should happen, but hopefully this will stop both
Profile and Editor API sentry being full of this error.

As far as I can tell, this call happens when a student first opens a
project (lesson). As they do not have their own remix yet, the fallback
gets the project, and that calls the `.with_user` that is causing this
problem.